Adapt watched_link.show view to encore
