+++ assets/js/watchedLink/chartInit.js
@@ -1,0 +1,68 @@
+
+// Colors
+var chartColors = {
+	red: 'rgb(255, 99, 132)',
+	orange: 'rgb(255, 159, 64)',
+	yellow: 'rgb(255, 205, 86)',
+	green: 'rgb(75, 192, 192)',
+	blue: 'rgb(54, 162, 235)',
+	purple: 'rgb(153, 102, 255)',
+	grey: 'rgb(201, 203, 207)'
+};
+ 
+// TO DO: responsive chart
+var ctx = document.getElementById('chart1').getContext('2d');
+ctx.canvas.width = 1000;
+ctx.canvas.height = 300;
+module.exports.ctx = ctx;
+
+
+module.exports.cfg = {
+    type: 'bar',
+    data: {
+        datasets: [{
+            label: 'Visits',
+            backgroundColor: chartColors.red,
+            borderColor: chartColors.red,
+            data: [],
+            type: 'line',
+            pointRadius: 0,
+            fill: false,
+            lineTension: 0,
+            borderWidth: 2
+        }]
+    },
+    options: {
+        scales: {
+            xAxes: [{
+                type: 'time',
+                distribution: 'series',
+                ticks: {
+                    source: 'data',
+                    autoSkip: true
+                }
+            }],
+            yAxes: [{
+                scaleLabel: {
+                    display: true,
+                    labelString: 'Visits'
+                }
+            }]
+        },
+        tooltips: {
+            intersect: false,
+            mode: 'index',
+            callbacks: {
+                label: function(tooltipItem, myData) {
+                    var label = myData.datasets[tooltipItem.datasetIndex].label || '';
+                    if (label) {
+                        label += ': ';
+                    }
+                    label += myData.datasets[tooltipItem.datasetIndex].data[tooltipItem.index].y;
+                    return label;
+                }
+            }
+        }
+    }
+};
+
