+++ templates/WatchedLink/show.html.twig
@@ -47,56 +47,14 @@
 		<button id="update" class="form-control">Update</button>
 		<br>
 
+		<input type="hidden" name="watched-link-id" id="watched-link-id" value="{{ WatchedLink.id }}">
+
 
     </article>
 
 {% endblock %}
 
 {% block javascripts %}
-	<script src="https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.22.2/moment.min.js"></script>
-	<script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/2.7.3/Chart.min.js"></script>
-	<script src="{{ asset("js/chartUtils.js") }}"></script>
-	<script src="{{ asset("js/chartInit.js") }}"></script>
-	
-	<script>	
-		$(function(){
-			// Get current watched link ID
-			var watchedLinkId = {{ WatchedLink.id }};
-
-			// Create dates
-			initStartDate = moment().subtract(6, "months").format(dateFormat);
-			initEndDate   = moment().add(1, "months").format(dateFormat);
-			
-			// Display right data on chart
-			getData(watchedLinkId, initStartDate , initEndDate, "m");
-			
-			// Filling inputs
-			$("#start-date").val(initStartDate);
-			$("#end-date").val(initEndDate);
-
-			// //
-			// Events
-			// //
-			$('#update').on('click', function() {
-
-				// Update type
-				var type = document.getElementById('type').value;
-				chart.config.data.datasets[0].type = type;
-				chart.update();
-
-				// Update data
-				startDate 	= $("#start-date").val();
-				endDate		= $("#end-date").val();
-				scale 		= $("#scale").val();
-				linkId 		= watchedLinkId;
-				getData(linkId, startDate, endDate, scale);
-			});
-
-			// Click to copy event
-			$(".click-to-copy").on('click', function(){
-				copyStringToClipboard($(this).text());	
-			});
-		});
-	</script>
+    {{ encore_entry_script_tags('watchedLink/show') }}
 {% endblock %}
