+++ webpack.config.js
@@ -25,6 +25,7 @@ Encore
      */
     .addEntry('app', './assets/js/app.js')
     .addEntry('watchedLink', './assets/js/watchedLink/index.js')
+    .addEntry('watchedLink/show', './assets/js/watchedLink/show.js')
     //.addEntry('page1', './assets/js/page1.js')
     //.addEntry('page2', './assets/js/page2.js')
